refactor(tagging): use canonical service module
