+++ tests/server/services/tagging/test_tagging_service.py
@@ -14,7 +14,7 @@ from reflexio.models.config_schema import (
     StorageConfigSQLite,
     UserPlaybookExtractorConfig,
 )
-from reflexio.server.services.tagging.tagging_service import TaggingService, TagsOutput
+from reflexio.server.services.tagging.service import TaggingService, TagsOutput
 
 
 class FakeStorage:
@@ -145,7 +145,7 @@ def make_config(
 def test_tagging_service_updates_profiles_and_playbooks(monkeypatch: Any) -> None:
     monkeypatch.delenv("MOCK_LLM_RESPONSE", raising=False)
     monkeypatch.setattr(
-        "reflexio.server.services.tagging.tagging_service.SiteVarManager.get_site_var",
+        "reflexio.server.services.tagging.service.SiteVarManager.get_site_var",
         lambda *_: {},
     )
     storage = FakeStorage()
@@ -183,7 +183,7 @@ def test_tagging_service_skips_entity_types_without_tagging_prompts(
 ) -> None:
     monkeypatch.delenv("MOCK_LLM_RESPONSE", raising=False)
     monkeypatch.setattr(
-        "reflexio.server.services.tagging.tagging_service.SiteVarManager.get_site_var",
+        "reflexio.server.services.tagging.service.SiteVarManager.get_site_var",
         lambda *_: {},
     )
     storage = FakeStorage()
@@ -204,7 +204,7 @@ def test_tagging_service_skips_entity_types_without_tagging_prompts(
 def test_tagging_service_can_scope_to_profiles_only(monkeypatch: Any) -> None:
     monkeypatch.delenv("MOCK_LLM_RESPONSE", raising=False)
     monkeypatch.setattr(
-        "reflexio.server.services.tagging.tagging_service.SiteVarManager.get_site_var",
+        "reflexio.server.services.tagging.service.SiteVarManager.get_site_var",
         lambda *_: {},
     )
     storage = FakeStorage()
@@ -226,7 +226,7 @@ def test_tagging_service_can_scope_to_profiles_only(monkeypatch: Any) -> None:
 def test_tagging_service_skips_already_tagged_entities(monkeypatch: Any) -> None:
     monkeypatch.delenv("MOCK_LLM_RESPONSE", raising=False)
     monkeypatch.setattr(
-        "reflexio.server.services.tagging.tagging_service.SiteVarManager.get_site_var",
+        "reflexio.server.services.tagging.service.SiteVarManager.get_site_var",
         lambda *_: {},
     )
     storage = FakeStorage()
